feat: Refactor getDailyReviewingRating as getReviewingRating to impl weekly, monthly features
