+++ src/main/java/com/somartreview/reviewmate/service/dashboard/DashboardService.java
@@ -1,10 +1,11 @@
-package com.somartreview.reviewmate.service;
+package com.somartreview.reviewmate.service.dashboard;
 
 import com.somartreview.reviewmate.domain.reservation.ReservationRepository;
 import com.somartreview.reviewmate.service.partners.PartnerCompanyService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.time.DayOfWeek;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Objects;
@@ -17,12 +18,11 @@ public class DashboardService {
     private final PartnerCompanyService partnerCompanyService;
 
 
-    public Float getDailyReviewingRate(String partnerDomain) {
+    public Float getReviewingRate(String partnerDomain, TimeSeriesUnit timeSeriesUnit) {
         partnerCompanyService.validateExistingPartnerDomain(partnerDomain);
 
-        LocalDateTime now = LocalDateTime.now();
-        LocalDateTime todayStart = LocalDateTime.of(now.getYear(), now.getMonth(), now.getDayOfMonth(), 0, 0, 0);
-        List<Long> reviewFks = reservationRepository.findAllReviewFKsByCreatedAtGreaterThanEqual(partnerDomain, todayStart);
+        LocalDateTime countingStartDateTime = getCountingStartDateTime(timeSeriesUnit);
+        List<Long> reviewFks = reservationRepository.findAllReviewFKsByCreatedAtGreaterThanEqual(partnerDomain, countingStartDateTime);
 
         long todayReservationCount = reviewFks.size();
         if (todayReservationCount == 0) {
@@ -31,5 +31,14 @@ public class DashboardService {
 
         long todayReviewCount = reviewFks.stream().filter(Objects::nonNull).count();
         return (float) todayReviewCount / todayReservationCount * 100;
+    }
+
+    private LocalDateTime getCountingStartDateTime(TimeSeriesUnit timeSeriesUnit) {
+        LocalDateTime now = LocalDateTime.now();
+        return switch (timeSeriesUnit) {
+            case DAILY -> LocalDateTime.of(now.getYear(), now.getMonth(), now.getDayOfMonth(), 0, 0, 0);
+            case WEEKLY -> now.with(DayOfWeek.MONDAY);
+            case MONTHLY -> LocalDateTime.of(now.getYear(), now.getMonth(), 1, 0, 0, 0);
+        };
     }
 }
