+++ src/main/java/com/somartreview/reviewmate/service/dashboard/TimeSeriesUnit.java
@@ -1,0 +1,7 @@
+package com.somartreview.reviewmate.service.dashboard;
+
+public enum TimeSeriesUnit {
+    DAILY,
+    WEEKLY,
+    MONTHLY
+}
